pyproject.toml
[project]
name = "grafos-trabalho"
version = "0.1.0"
description = ""
authors = [
    {name = "Jfdeev",email = "jfrsoares161205@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain (==0.3.26)",
    "streamlit (>=1.49.1,<2.0.0)",
    "langgraph (>=0.6.7,<0.7.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-ollama (>=0.3.8,<0.4.0)",
    "tavily-python (>=0.7.12,<0.8.0)",
    "networkx (>=3.0,<4.0.0)",
    "plotly (>=5.17.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
